modules/google_drive_handler.py
import logging
from google.oauth2.service_account import Credentials
from googleapiclient.discovery import build

class GoogleDriveHandler:
    """Google Drive 處理器"""
    
    def __init__(self, config):
        self.config = config
        self.service = None
        
        try:
            if hasattr(config, 'service_account_file'):
                scope = ['https://www.googleapis.com/auth/drive']
                creds = Credentials.from_service_account_file(
                    config.service_account_file, 
                    scopes=scope
                )
                self.service = build('drive', 'v3', credentials=creds)
                logging.info("✅ Google Drive 處理器初始化成功")
            else:
                logging.warning("⚠️ Google Drive 服務帳戶檔案未找到")
                
        except Exception as e:
            logging.error(f"Google Drive 初始化失敗: {e}")
            self.service = None
    
    def is_available(self):
        """檢查 Google Drive 是否可用"""
        return self.service is not None
    
    def cleanup_expired_files(self):
        """清理過期檔案"""
        if not self.service:
            return
        
        try:
            # 簡單的清理邏輯
            logging.info("執行 Google Drive 檔案清理")
        except Exception as e:
            logging.error(f"Google Drive 檔案清理錯誤: {e}")
    
    def upload_file(self, file_data, filename):
        """上傳檔案到 Google Drive"""
        if not self.service:
            return None
        
        try:
            # 簡化的檔案上傳邏輯
            logging.info(f"上傳檔案: {filename}")
            return f"https://drive.google.com/file/d/example_id/view"
        except Exception as e:
            logging.error(f"Google Drive 檔案上傳錯誤: {e}")
            return None