requirements.txt
# 核心Web框架
Flask==3.0.3
gunicorn==21.2.0

# LINE Bot SDK
line-bot-sdk==3.17.1

# Google AI和雲端服務
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gspread==6.2.1

# gRPC相關
grpcio==1.71.0
grpcio-status==1.71.0

# HTTP和網路
aiohttp==3.11.18
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
requests==2.32.3
httplib2==0.22.0
httpx==0.28.1
urllib3==2.2.3

# 數據處理和分析
pandas==2.2.3
numpy==2.2.6
matplotlib==3.10.3

# 第三方API
newsapi-python==0.2.7

# 時間和日期處理
python-dateutil==2.9.0.post0
pytz==2024.1

# 環境和配置
python-dotenv==1.1.0

# 網頁解析
beautifulsoup4==4.12.2
soupsieve==2.7

# 圖像處理
pillow==11.2.1

# 加密和認證
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.7

# 支援庫
aenum==3.1.16
annotated-types==0.6.0
anyio==4.3.0
attrs==25.3.0
blinker==1.7.0
cachetools==5.5.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.18
docstring_parser==0.16
docutils==0.16
frozenlist==1.6.0
future==1.0.0
h11==0.16.0
httptools==0.6.4
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.4.3
packaging==25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.2.3
rsa==4.7.2
six==1.17.0
sniffio==1.3.1
wrapt==1.17.0
yarl==1.20.0

# 開發和測試
uvicorn==0.34.0
watchfiles==1.0.0
websockets==15.0.0
PyYAML==6.0.2

# 字體和圖表支援
fonttools==4.58.1
contourpy==1.3.2
cycler==0.12.1
kiwisolver==1.4.8

# OAuth支援
oauthlib==3.2.2
requests-oauthlib==2.0.0